+++ obj/Debug/netstandard2.1/MonitorLabels.AssemblyInfo.cs
@@ -15,7 +15,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyDescriptionAttribute("Adds a label to the icons on the monitor for players, radar boosters, enemies and" +
     " scrap")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.1.1.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.1.1+65dfafe0c7e82ea8e7113386612a713b49fd8219")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.1.1+8238f056026aad5743ad23ff59903d197f26d4e0")]
 [assembly: System.Reflection.AssemblyProductAttribute("MonitorLabels")]
 [assembly: System.Reflection.AssemblyTitleAttribute("MonitorLabels")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.1.1.0")]